refactor(keyboard): rename 'Keyboard' to 'identifiers'
