+++ packages/retail-ui/components/CurrencyInput/CurrencyInputKeyboardActions.tsx
@@ -1,4 +1,4 @@
-import * as Keyboard from '../../lib/events/keyboard/Keyboard';
+import * as Keyboard from '../../lib/events/keyboard/identifiers';
 import { KeyboardActionExctracterBuilder } from '../internal/extractKeyboardAction';
 
 export const CURRENCY_INPUT_ACTIONS = {
