+++ packages/retail-ui/components/DateInput/helpers/DateInputKeyboardActions.tsx
@@ -1,7 +1,7 @@
 // @ts-ignore noUnusedVar
 import * as React from 'react';
 import { SEPARATOR } from '../../../lib/date/constants';
-import * as Keyboard from '../../../lib/events/keyboard/Keyboard';
+import * as Keyboard from '../../../lib/events/keyboard/identifiers';
 import { KeyboardActionExctracterBuilder } from '../../internal/extractKeyboardAction';
 
 const delimiters = [
