+++ packages/retail-ui/components/Switcher/Switcher.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import * as PropTypes from 'prop-types';
-import { isKeyArrowHorizontal, isKeyArrowLeft, isKeyEnter } from '../../lib/events/keyboard/Keyboard';
+import { isKeyArrowHorizontal, isKeyArrowLeft, isKeyEnter } from '../../lib/events/keyboard/identifiers';
 import Group from '../Group';
 import Button, { ButtonSize } from '../Button';
 import styles from './Switcher.module.less';
